+++ src/lib.rs
@@ -2,6 +2,7 @@
 
 pub mod asteroid;
 pub mod graphics;
+pub mod player;
 pub mod simulation;
 pub mod spatial_partition;
 pub mod testing;